phone_agent/requirements.txt
flask
twilio
python-dotenv
pandas
pydub
deepgram-sdk
google-generativeai
flask_cors
